integration commitment visualize
